backend/requirements-cloudrun.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
requests>=2.31.0
aiohttp>=3.8.0
google-cloud-firestore>=2.14.0
emergentintegrations --extra-index-url https://d33sy5i8bnduwe.cloudfront.net/simple/
